requirements/common.txt
Django==1.7
djangorestframework
bitcasa